WebsiteCorona/backend/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
Django==3.0.4
django-celery==3.3.1
django-cors-headers==3.2.1
django-crontab==0.7.1
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
Fiona==1.8.13.post1
geopandas==0.7.0
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
kombu==3.0.37
lxml==4.5.0
munch==2.5.0
numpy==1.18.2
pandas==1.0.3
phonenumbers==8.12.1
pkg-resources==0.0.0
pycountry==19.8.18
pyproj==2.6.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
selenium==3.141.0
Shapely==1.7.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
zipp==3.1.0
